add [to/from]EncodingQuery, [to/from]FieldQuery, methods
